Huge performance improvements
It was time for an optimisation round, and all those syscalls writing
out to the terminal were bad for performance. All output styles now
write to an intermedia buffer, which is written out from the main
program -- which also means it's now not necessary to run those scripts
on a terminal. That could be useful, who knows.